kingdom builder v2 prep
