+++ src/module/apps/attribute-builder/components/remaining-indicator.svelte
